KrishiVaani2/requirement.txt
streamlit==1.38.0
transformers==4.44.2
torch==2.4.1
langchain==0.3.0
faiss-cpu==1.8.0
sentence-transformers==3.0.1
requests==2.32.3
Pillow==10.4.0
numpy==2.1.1
indic-nlp-library==0.92  # For Indic language processing